# swagger.yaml
openapi: 3.0.3
info:
  title: MeetFlow API
  version: 1.0.0
  description: |
    ## MeetFlow Backend REST API

    This API powers the MeetFlow SaaS platform. It is divided into four
    top-level sections:

    | Section          | Base Path          | Description                                    |
    |------------------|--------------------|------------------------------------------------|
    | **Common**       | `/api/v1/auth` & `/api/v1/account`          | Authentication & account management            |
    | **Master**       | `/api/v1/master`   | MeetFlow internal admin panel (master context) |
    | **Organization** | `/api/v1/organization` | Organisation-level management (coming soon) |
    | **Workspace**    | `/api/v1/workspace`    | Workspace & channel management (coming soon)|

    ### Authentication
    All **Master**, **Organization**, and **Workspace** routes require a valid
    JWT `accessToken` in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```
    Use the **Login** endpoint to obtain tokens.

    ### Standard Response Envelope
    Every response (success or error) is wrapped in:
    ```json
    {
      "status": {
        "response_code": 200,
        "response_message": "Success!"
      },
      "message": "Human-readable message",
      "data": { ... }
    }
    ```
    Error responses replace `data` with an `error` object containing `details`.

  contact:
    name: MeetFlow Engineering
  license:
    name: ISC
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token returned by POST /auth/login.
        Valid for 1 day (or 30 days when rememberMe is true).
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: Numeric record ID (BIGINT primary key)
      schema:
        type: integer
        format: int64
        example: 1
    QuerySearch:
      name: search
      in: query
      required: false
      description: Full-text search / keyword filter
      schema:
        type: string
        example: admin
    QueryPage:
      name: page
      in: query
      required: false
      description: Page number for pagination (1-based)
      schema:
        type: integer
        default: 1
        minimum: 1
    QueryLimit:
      name: limit
      in: query
      required: false
      description: Number of records per page
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
    QueryIsActive:
      name: is_active
      in: query
      required: false
      description: Filter by active / inactive status
      schema:
        type: boolean
  schemas:
...
    StatusBlock:
      type: object
      properties:
        response_code:
          type: integer
          example: 200
        response_message:
          type: string
          example: "Success!"
    SuccessEnvelope:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/StatusBlock"
        message:
          type: string
          example: "Operation successful"
        data:
          description: Payload — varies per endpoint
          nullable: true
    ErrorEnvelope:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/StatusBlock"
        message:
          type: string
          example: "Something went wrong!"
        error:
          type: object
          properties:
            details:
              type: string
              example: "Invalid credentials."
    DeletedResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              nullable: true
              example: null

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    MasterUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@gmail.com
        role_id:
          type: integer
          example: 2
        is_active:
          type: boolean
          example: true
        last_login:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required:
        - first_name
        - last_name
        - email
        - password
        - role_id
      properties:
        first_name:
          type: string
          maxLength: 50
          example: John
        last_name:
          type: string
          maxLength: 50
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@gmail.com
        password:
          type: string
          format: password
          minLength: 8
          example: "StrongPass@123"
        role_id:
          type: integer
          example: 2
        is_active:
          type: boolean
          default: true
    UpdateUserRequest:
      type: object
      properties:
        first_name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
        role_id:
          type: integer
        is_active:
          type: boolean

    # ── Plan ──────────────────────────────────────────────────────────────────
...
    BulkCreateUserItem:
      type: object
      required:
        - first_name
        - last_name
        - email
        - password
        - role_id
      properties:
        first_name:
          type: string
          maxLength: 50
          example: John
        last_name:
          type: string
          maxLength: 50
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@gmail.com
        password:
          type: string
          format: password
          minLength: 8
          example: "StrongPass@123"
        role_id:
          type: integer
          example: 2
        is_active:
          type: boolean
          default: true
    BulkCreateUserRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/BulkCreateUserItem"
    BulkUpdateUserItem:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          example: 1
        first_name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
        role_id:
          type: integer
        is_active:
          type: boolean
    BulkUpdateUserRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/BulkUpdateUserItem"
    BulkDeleteUserRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: integer
            format: int64
          example: [1, 2, 3]

    # ── Bulk Plans ────────────────────────────────────────────────────────────
...
  responses:
    UnauthorizedError:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            status:
              response_code: 401
              response_message: "Something went wrong!"
            message: "Access token is required."
    ForbiddenError:
      description: Authenticated but insufficient permissions (wrong context)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            status:
              response_code: 403
              response_message: "Something went wrong!"
            message: "This route is only accessible to master users."
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequestError:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
paths:
...
  /master/users/get-all-users:
    get:
      tags: [Master / Users]
      summary: Get all master users
      operationId: getAllUsers
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QuerySearch"
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryIsActive"
        - name: role_id
          in: query
          required: false
          schema:
            type: integer
            example: 2
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/MasterUser"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /master/users/get-user-by-id/{id}:
    get:
      tags: [Master / Users]
      summary: Get master user by ID
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MasterUser"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/users/create-user:
    post:
      tags: [Master / Users]
      summary: Create a master user
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            example:
              first_name: Jane
              last_name: Smith
              email: jane.smith@gmail.com
              password: "StrongPass@123"
              role_id: 2
              is_active: true
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MasterUser"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /master/users/update-user/{id}:
    patch:
      tags: [Master / Users]
      summary: Update a master user
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MasterUser"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/users/delete-user/{id}:
    delete:
      tags: [Master / Users]
      summary: Soft delete a master user
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ─── Bulk User Operations ─────────────────────────────────────────────────
  /master/users/bulk-create:
    post:
      tags: [Master / Users]
      summary: Bulk create master users
      description: |
        Create multiple MeetFlow internal staff accounts in a single request.
        **Limits:** Maximum 100 items. Each email must be unique.
      operationId: bulkCreateUsers
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreateUserRequest"
      responses:
        "200":
          description: Bulk create completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          successful:
                            type: array
                            items:
                              $ref: "#/components/schemas/MasterUser"
                          failed:
                            type: array
                          summary:
                            type: object
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /master/users/bulk-update:
    patch:
      tags: [Master / Users]
      summary: Bulk update master users
      description: |
        Update multiple MeetFlow internal staff accounts in a single request.
        Each item must include an `id`.
        **Limits:** Maximum 100 items.
      operationId: bulkUpdateUsers
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpdateUserRequest"
      responses:
        "200":
          description: Bulk update completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /master/users/bulk-delete:
    delete:
      tags: [Master / Users]
      summary: Bulk delete master users
      description: |
        Soft delete multiple MeetFlow internal staff accounts by ID.
        **Limits:** Maximum 100 IDs.
      operationId: bulkDeleteUsers
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteUserRequest"
      responses:
        "200":
          description: Bulk delete completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

  # ═══════════════════════════════════════════════════════════════════════════
  # MASTER — Plans
  # ═══════════════════════════════════════════════════════════════════════════
...